Updated Home screen and other improvments.
Updated the main menu to have the environment. You can quit, play or reset data, this code is to be deeloped further in order to work with the appropriate stats. The respawn counter has been fixed to only allow adding once per x seconds, there is a possibiltiy to streamilne the code by not referancing playerprefs every time you want to add to it but mae a local variable instead(may be unnessacery).